+++ pyage/screens/menu/__init__.py
@@ -1,22 +1,25 @@
 """
-Any pre-defined items for the :class:`pyage.screens.Menu` screen can be found here.
+A menu screen as well as its items can be found here.
 
-.. currentmodule:: pyage.screens.items
+.. currentmodule:: pyage.screens.menu
 
 .. autosummary::
    :toctree:
 
    Button
+   Menu
    MenuItem
    TextInput
 """
 
 from .button import Button
+from .menu import Menu
 from .menu_item import MenuItem
 from .text_input import TextInput
 
 __all__ = (
     "Button",
+    "Menu",
     "MenuItem",
     "TextInput",
 )